refactor: 移动 use-spring-resume-on-visible 到 utils 目录
- 将 hook 从组件目录移动到公共 utils 目录
- 更新所有相关引用路径
- 保持原有功能不变
