+++ src/components/center-popup/center-popup.tsx
@@ -8,7 +8,7 @@ import { renderToContainer } from '../../utils/render-to-container'
 import { ShouldRender } from '../../utils/should-render'
 import { useInnerVisible } from '../../utils/use-inner-visible'
 import { useLockScroll } from '../../utils/use-lock-scroll'
-import { useSpringResumeOnVisible } from '../popup/use-spring-resume-on-visible'
+import { useSpringResumeOnVisible } from '../../utils/use-spring-resume-on-visible'
 import { mergeProps } from '../../utils/with-default-props'
 import { withStopPropagation } from '../../utils/with-stop-propagation'
 import { useConfig } from '../config-provider'
